refactor: flatten uphoster_utils subdirectories into uploaders_utils
Move buzzheavier_utils/upload.py, gofile_utils/upload.py,
pixeldrain_utils/upload.py into a single uploaders_utils/ directory
with descriptive names. Each subdirectory had only one file.
