+++ bot/helper/mirror_leech_utils/uphoster_utils/multi_upload.py
@@ -1,13 +1,13 @@
 from asyncio import gather
 from logging import getLogger
 
-from bot.helper.mirror_leech_utils.uphoster_utils.buzzheavier_utils.upload import (
+from bot.helper.mirror_leech_utils.uphoster_utils.uploaders_utils.buzzheavier_uploader import (
     BuzzHeavierUpload,
 )
-from bot.helper.mirror_leech_utils.uphoster_utils.gofile_utils.upload import (
+from bot.helper.mirror_leech_utils.uphoster_utils.uploaders_utils.gofile_uploader import (
     GoFileUpload,
 )
-from bot.helper.mirror_leech_utils.uphoster_utils.pixeldrain_utils.upload import (
+from bot.helper.mirror_leech_utils.uphoster_utils.uploaders_utils.pixeldrain_uploader import (
     PixelDrainUpload,
 )
